requirements.txt
streamlit
numpy
pandas
scikit-learn
matplotlib
plotly
PyPDF2
python-docx
reportlab
Pillow
beautifulsoup4
openai
